## Addendum (5/31/2023)
* Make --enable-threadsafe default to disabled.
* Make threadsafe work with automake on MinGW and Cygwin.
* Add threadsafe testing to Ubuntu, MinGW, and Cygwin github actions.
* Cleanup the library search for PTHREADS4W in CMakeLists.txt.
* Add a test in *nc_test* -- run_thread_nc_test.sh and thread_nc_test.c -- that modifies the *nc_test.c* to use multi-threading.
* Add a new installed header: *netcdf_threadsafe.h*.
* Rename some internal header and .c files.
* Re-enable DAP4.
* Remove "const" attribute from the dispatch table pointers because it fails under multi-threading.
* Unify definitions of e.g. nulldup.
* Fix some potential memory leaks.

+++ nc_test/run_thread_nc_test.sh
@@ -5,16 +5,29 @@ if test "x$srcdir" = x ; then srcdir=`pwd`; fi
 
 set -e
 
+NTHREADS=4
+NMSGS=4
+
+isolate "testdir_nc_tests_threads"
+
+# Move into test directory
+cd $ISOPATH
+
+if test 1 = 1 ; then
 echo "*** Testing netcdf-3 thread safe execution"
-${execdir}/tst_threads 3 "tmp_threadsafe_3_%d.nc"
+${execdir}/thread_nc_test -T ${NTHREADS} -M ${NMSGS}
 echo "PASS: thread safe netcdf-3"
 rm -f tmp_threadsafe*.nc
+fi
 
+if test 1 = 1 ; then
 if test "x$FEATURE_HDF5" = xyes ; then
 echo "*** Testing netcdf-4 thread safe execution"
-${execdir}/tst_threads 4 "tmp_threadsafe_4_%d.nc"
+#export NETCDF_LOG_LEVEL=5
+${execdir}/thread_nc_test -T ${NTHREADS} -M ${NMSGS}
 echo "PASS: thread safe netcdf-4"
 fi
 rm -f tmp_threadsafe*.nc
+fi
 
 exit
